Replace cherrypy with uwsgi
