+++ coverart_redirect/wsgi.py
@@ -1,11 +1,17 @@
-# Copyright (C) 2011 Lukas Lalinsky
+# Copyright (C) 2017 MetaBrainz Foundation
 # Distributed under the MIT license, see the LICENSE file for details.
 
-# Simple WSGI module intended to be used by uWSGI, e.g.:
-# uwsgi -w acoustid.wsgi --pythonpath ~/acoustid/ --env COVERART_REDIRECT_CONFIG=~/acoustid/acoustid.conf --http :9090
-# uwsgi -w acoustid.wsgi --pythonpath ~/acoustid/ --env COVERART_REDITECT_CONFIG=~/acoustid/acoustid.conf -M -L --socket 127.0.0.1:1717
+# Simple WSGI module intended to be used by uWSGI.
 
-import os
-from coverart_redirect.server import make_application
+from coverart_redirect.server import Server
+from coverart_redirect.config import load_config
+from coverart_redirect.loggers import init_raven_client
 
-application = make_application(os.environ['COVERART_REDIRECT_CONFIG'])
+
+config = load_config()
+
+sentry_dsn = config.sentry.dsn
+if sentry_dsn:
+    init_raven_client(sentry_dsn)
+
+application = Server(config)
