+++ docker/prod/redirect.service
@@ -1,3 +1,3 @@
 #!/bin/bash
-export PYTHONPATH="/code/"
-exec python /code/coverart_redirect_server.py
+
+exec uwsgi /etc/uwsgi/uwsgi.ini
