+++ docker/prod/uwsgi.ini
@@ -1,0 +1,9 @@
+[uwsgi]
+master = true
+http = 0.0.0.0:8080
+module = coverart_redirect.wsgi
+callable = application
+chdir = /code/
+enable-threads = true
+processes = 10
+die-on-term = true
