+++ requirements.txt
@@ -1,4 +1,3 @@
-cherrypy==7.0.0
 coverage==4.0.3
 pytest==3.0.4
 pytest-cov==2.4.0
@@ -7,3 +6,4 @@ raven==6.1.0
 requests==2.18.3
 sqlalchemy==1.1.15
 werkzeug==0.11.2
+uWSGI==2.0.15
